// src/generated/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type Movies = {
  __typename?: 'Movies';
  age?: Maybe<Scalars['Int']>;
  category?: Maybe<Scalars['String']>;
  name?: Maybe<Scalars['String']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  addUser?: Maybe<Scalars['String']>;
};


export type MutationAddUserArgs = {
  input: UserInput;
};

export type Query = {
  __typename?: 'Query';
  getMovies?: Maybe<Array<Maybe<Movies>>>;
  getUsers?: Maybe<Array<Maybe<Users>>>;
};

export type UserInput = {
  email: Scalars['String'];
  lastname: Scalars['String'];
  name: Scalars['String'];
  password: Scalars['String'];
};